Refactor user ID handling to use Long instead of Integer across multiple services and repositories; add employee monthly leave and task count endpoints

+++ src/main/java/org/ems/demo/controller/ImageController.java
@@ -21,7 +21,7 @@ public class ImageController {
     @PermissionRequired(values = {"ROLE_USER", "ROLE_ADMIN", "ROLE_EMP"})
     public ResponseEntity<SuccessResponse> uploadProfile(
             @RequestParam("file") MultipartFile file,
-            @RequestParam("userId") Integer userId
+            @RequestParam("userId") Long userId
     ) {
 
         SuccessResponse successResponse = SuccessResponse.builder()
