+++ src/main/java/org/ems/demo/repository/impl/UserNativeRepositoryImpl.java
@@ -35,7 +35,7 @@ public class UserNativeRepositoryImpl implements UserNativeRepository {
                             rs.getString("c.logo")
                     );
                     return new UserEntity(
-                            rs.getInt("u.id"),
+                            rs.getLong("u.id"),
                             rs.getString("u.first_name"),
                             rs.getString("u.last_name"),
                             rs.getString("u.email"),
